http/Client: Fix connectUnix compilation
`connectUnix()` was broken by the changes to how we manage the
connection pool that were made when length tracking was removed
from `std.DoublyLinkedList`. If a developer attempted to use it
currently, it would fail to compile due to it relying on incorrect
types.

This commit fixes all compilation errors in this function, bringing
it in line with `connectTcp()`.