pwhash: Make each hashing function unique
There's no real reason for these to be generic functions - the only
reason they're written as such is for the benchmarking code. Otherwise,
being generic harms their usability and contorts these functions in
a way that violates Zig Zen.

This commit rewrites the three pwhash functions and their associated
structs to be non-generic, allowing for each API to properly communicate
their intended inputs in their function arguments and type fields.