Module: omit frame pointer in ReleaseFast
Because it frees up a register it can have performance implications on
the program and because we're optimizing for speed and not debugging,
I think it's a no-brainer to make this true.
In general, I expect performance with this change to be strictly better or unchanged,
both at the cost of it being more difficult to obtain a stack trace but
that's not the goal of ReleaseFast.