rename: Add error.CircularLoop
Adds error.CircularLoop to rename functions.

According to the man page for rename, "EINVAL The old pathname names
an ancestor directory of the new pathname, or either pathname argument
contains a final component that is dot or dot-dot." The later is always
programmer error, so returning error.CircularLoop should always be
correct.

Based on empirical testing, SHARING_VIOLATION is returned by Windows to
indicate a circular loop.