rename: Add error.CircularLoop
Adds error.CircularLoop to rename functions.

For POSIX systems: according to the man page for rename, "EINVAL The
old pathname names an ancestor directory of the new pathname, or either
pathname argument contains a final component that is dot or dot-dot."
The later is always programmer error, so returning error.CircularLoop
should always be correct.

For WASI systems: the error should be the same as POSIX. The
documentation for INVAL says "Invalid argument, similar to `EINVAL` in
POSIX." This holds true for wasmtime 33.0.0 on my system, however the
zig CI returns PERM, so the test is skipped on WASI.

For Windows systems: based on empirical testing, SHARING_VIOLATION is
returned on Windows 10 to indicate a circular loop. At some point on
Windows 11, INVALID_PARAMATER started being used.