wasm2c warm.h: make panic a macro to propagate noreturn
I got the following warning when compiling zig-wasm2c:

C:\git\zig\stage1\wasm.h(55) : warning C4715: 'WasmMut_toC': not all control paths return a value

This is because the C compiler doesn't understand that the panic
function from `panic.h` is noreturn.  I addressed this making panic
a macro so the caller sees that abort is called and is able to see
that it's a "noreturn codepath".