rename: Add and document errors caused by circular loops
On POSIX systems, EINVAL is used to indicate circular loops;
additionally, corresponding with the use of the error in other
filesystem syscalls, it is used to indicate bad file names.

Based on empirical testing on Windows 10, SHARING_VIOLATION is
used to indicate a circular loop. On Windows 11, INVALID_PARAMATER
is used instead.